PWN-3206 - ssl cert added and used in app (#509)
* PWN-3206 - ssl cert added and used in app

* PWN-3206 - reformat code

* PWN-3206 - add to feature flags

* PWN-3206 - add SSL certificate to CircleCI

* PWN-3206 - add SSL certificate to CircleCI #2

* PWN-3206 - small renaming

* PWN-3206 - remove pining enabled on prod

* PWN-3206 - changes according to commnets

* PWN-3206 - CertificateManager.kt extracted
