+++ app/src/main/java/org/p2p/wallet/infrastructure/network/NetworkModule.kt
@@ -11,7 +11,6 @@ import org.koin.core.scope.Scope
 import org.koin.dsl.module
 import org.p2p.wallet.BuildConfig
 import org.p2p.wallet.R
-import org.p2p.wallet.common.AppFeatureFlags
 import org.p2p.wallet.common.crashlytics.CrashHttpLoggingInterceptor
 import org.p2p.wallet.common.di.InjectionModule
 import org.p2p.wallet.home.HomeModule.MOONPAY_QUALIFIER
@@ -23,35 +22,25 @@ import org.p2p.wallet.infrastructure.network.interceptor.MoonpayErrorInterceptor
 import org.p2p.wallet.infrastructure.network.interceptor.RpcInterceptor
 import org.p2p.wallet.infrastructure.network.interceptor.RpcSolanaInterceptor
 import org.p2p.wallet.infrastructure.network.provider.TokenKeyProvider
+import org.p2p.wallet.infrastructure.network.ssl.CertificateManager
 import org.p2p.wallet.push_notifications.PushNotificationsModule.NOTIFICATION_SERVICE_RETROFIT_QUALIFIER
 import org.p2p.wallet.rpc.RpcModule.RPC_RETROFIT_QUALIFIER
 import org.p2p.wallet.rpc.RpcModule.RPC_SOLANA_RETROFIT_QUALIFIER
 import org.p2p.wallet.updates.ConnectionStateProvider
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import timber.log.Timber
-import java.io.InputStream
 import java.math.BigDecimal
-import java.security.GeneralSecurityException
-import java.security.KeyStore
-import java.security.cert.Certificate
-import java.security.cert.CertificateFactory
 import java.util.concurrent.TimeUnit
-import javax.net.ssl.SSLContext
-import javax.net.ssl.TrustManager
-import javax.net.ssl.TrustManagerFactory
-import javax.net.ssl.X509TrustManager
 
 object NetworkModule : InjectionModule {
 
     const val DEFAULT_CONNECT_TIMEOUT_SECONDS = 30L
     const val DEFAULT_READ_TIMEOUT_SECONDS = 30L
 
-    private const val SSL_CERT_TAG = "SSL_CERT"
-
     override fun create() = module {
         single { EnvironmentManager(get(), get()) }
         single { TokenKeyProvider(get()) }
+        single { CertificateManager(get(), get()) }
 
         single {
             GsonBuilder()
@@ -87,7 +76,12 @@ object NetworkModule : InjectionModule {
         single(named(RPC_SOLANA_RETROFIT_QUALIFIER)) {
             val environment = get<EnvironmentManager>().loadRpcEnvironment()
             val rpcApiUrl = environment.endpoint
-            getRetrofit(rpcApiUrl, "RpcSolana", RpcSolanaInterceptor(get()))
+            getRetrofit(
+                baseUrl = rpcApiUrl,
+                tag = "RpcSolana",
+                resources = get(),
+                interceptor = RpcSolanaInterceptor(get())
+            )
         }
 
         single(named(NOTIFICATION_SERVICE_RETROFIT_QUALIFIER)) {
@@ -107,7 +101,6 @@ object NetworkModule : InjectionModule {
         resources: Resources,
         interceptor: Interceptor?
     ): Retrofit {
-
         return Retrofit.Builder()
             .baseUrl(baseUrl)
             .addConverterFactory(GsonConverterFactory.create(get()))
@@ -126,21 +119,8 @@ object NetworkModule : InjectionModule {
             .readTimeout(DEFAULT_CONNECT_TIMEOUT_SECONDS, TimeUnit.SECONDS)
             .connectTimeout(DEFAULT_READ_TIMEOUT_SECONDS, TimeUnit.SECONDS)
             .apply {
-                val appFeatureFlags: AppFeatureFlags = get()
-                if (appFeatureFlags.isSslPinningEnabled) {
-                    try {
-                        createCertificate(resources.openRawResource(R.raw.cert)).apply {
-                            systemDefaultTrustManager()?.let { trustManager ->
-                                Timber.tag(SSL_CERT_TAG).i("SslSocketFactory successfully added with cert")
-                                sslSocketFactory(socketFactory, trustManager)
-                            }
-                        }
-                    } catch (e: Exception) {
-                        if (!BuildConfig.DEBUG) {
-                            Timber.tag(SSL_CERT_TAG).e(e, "Error on opening SSL cert")
-                        }
-                    }
-                }
+                val certificateManager: CertificateManager = get()
+                certificateManager.setCertificate(this)
                 if (BuildConfig.CRASHLYTICS_ENABLED) {
                     addInterceptor(CrashHttpLoggingInterceptor())
                 }
@@ -155,44 +135,5 @@ object NetworkModule : InjectionModule {
             }
             .addNetworkInterceptor(ContentTypeInterceptor())
             .build()
-    }
-
-    private fun createCertificate(trustedCertificateIS: InputStream): SSLContext {
-        val certificateFactory: CertificateFactory = CertificateFactory.getInstance("X.509")
-        val certificate: Certificate = trustedCertificateIS.use { trustedCertificate ->
-            certificateFactory.generateCertificate(trustedCertificate)
-        }
-
-        // creating a KeyStore containing our trusted CAs
-        val keyStoreType: String = KeyStore.getDefaultType()
-        val keyStore: KeyStore = KeyStore.getInstance(keyStoreType)
-        keyStore.load(null, null)
-        keyStore.setCertificateEntry("ca", certificate)
-
-        // creating a TrustManager that trusts the CAs in our KeyStore
-        val trustAlgorithm: String = TrustManagerFactory.getDefaultAlgorithm()
-        val trustManagerFactory: TrustManagerFactory = TrustManagerFactory.getInstance(trustAlgorithm)
-        trustManagerFactory.init(keyStore)
-
-        // creating an SSLSocketFactory that uses our TrustManager
-        val sslContext: SSLContext = SSLContext.getInstance("TLS")
-        sslContext.init(null, trustManagerFactory.trustManagers, null)
-        return sslContext
-    }
-
-    private fun systemDefaultTrustManager(): X509TrustManager? {
-        return try {
-            val trustManagerFactory: TrustManagerFactory =
-                TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm())
-            trustManagerFactory.init(null as KeyStore?)
-            val trustManagers: Array<TrustManager> = trustManagerFactory.trustManagers
-            check(!(trustManagers.size != 1 || trustManagers[0] !is X509TrustManager)) {
-                "Unexpected default trust managers:" + trustManagers.contentToString()
-            }
-            trustManagers[0] as X509TrustManager
-        } catch (e: GeneralSecurityException) {
-            // The system has no TLS. Just give up.
-            throw AssertionError()
-        }
     }
 }
